+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.AADEntitlementManagementConnectedOrganization.Tests.ps1
@@ -58,6 +58,10 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 return @(
                     @{
                         Id = '12345678-1234-1234-1234-123456789012'
+                        AdditionalProperties = @{
+                            '@odata.type' = '#microsoft.graph.user'
+                            userPrincipalName = 'John.smith@contoso.com'
+                        }
                     }
                 )
             }
@@ -65,6 +69,10 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 return @(
                     @{
                         Id = '12345678-1234-1234-1234-123456789012'
+                        AdditionalProperties = @{
+                            '@odata.type' = '#microsoft.graph.user'
+                            userPrincipalName = 'John.smith@contoso.com'
+                        }
                     }
                 )
             }
@@ -154,7 +162,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 $testParams = @{
                     Description      = 'ConnectedOrganization_Description'
                     DisplayName      = 'ConnectedOrganization_DisplayName'
-                    ExternalSponsors = @('12345678-1234-1234-1234-123456789012')
+                    ExternalSponsors = @('John.smith@contoso.com')
                     Id               = 'ConnectedOrganization_Id'
                     IdentitySources  = @(
                         (New-CimInstance -ClassName MSFT_AADEntitlementManagementConnectedOrganizationIdentitySource -Property @{
@@ -163,7 +171,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                             displayName      = 'IdentitySource_DisplayName'
                         } -ClientOnly)
                     )
-                    InternalSponsors = @('12345678-1234-1234-1234-123456789012')
+                    InternalSponsors = @('John.smith@contoso.com')
                     State            = 'configured'
                     Ensure           = 'Present'
                     Credential       = $Credential
@@ -198,7 +206,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 $testParams = @{
                     Description      = 'ConnectedOrganization_Description'
                     DisplayName      = 'ConnectedOrganization_DisplayName'
-                    ExternalSponsors = @('12345678-1234-1234-1234-123456789012')
+                    ExternalSponsors = @('John.smith@contoso.com')
                     Id               = 'ConnectedOrganization_Id'
                     IdentitySources  = @(
                         (New-CimInstance -ClassName MSFT_AADEntitlementManagementConnectedOrganizationIdentitySource -Property @{
@@ -207,7 +215,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                             displayName      = 'IdentitySource_DisplayName'
                         } -ClientOnly)
                     )
-                    InternalSponsors = @('12345678-1234-1234-1234-123456789012')
+                    InternalSponsors = @('John.smith@contoso.com')
                     State            = 'configured'
                     Ensure           = 'Absent'
                     Credential       = $Credential
@@ -276,9 +284,17 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     return @(
                         @{
                             Id = '12345678-1234-1234-1234-123456789012'
+                            AdditionalProperties = @{
+                                '@odata.type' = '#microsoft.graph.user'
+                                userPrincipalName = 'John.Smith@contoso.com'
+                            }
                         },
                         @{
                             Id = '12345678-1234-1234-1234-234567890123' #Drift
+                            AdditionalProperties = @{
+                                '@odata.type' = '#microsoft.graph.user'
+                                userPrincipalName = 'John.doe@contoso.com'
+                            }
                         }
                     )
                 }